WorkflowServer/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# AI / LLM
openai==1.30.1
google-generativeai==0.5.4

# Vector store
chromadb==0.4.24

# NumPy pinned below 2.0 for compatibility
numpy==1.26.4

# PDF extraction
pymupdf==1.24.3

# Web search
google-search-results==2.4.2
httpx==0.27.0

# Env / config
python-dotenv==1.0.1
pydantic-settings==2.2.1

# Utils
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]
cryptography